+++ checkout-ui-custom/checkout6-custom.js
@@ -1,1 +1,1 @@
-window.location.href = '/checkout-b2b'
+window.location.assign('/checkout-b2b')