+++ react/typings/index.ts
@@ -56,6 +56,7 @@ export type PaymentsBody = Array<
     installmentsInterestRate: number
     installmentsValue: number
     referenceValue: number | null
+    isBillingAddressDifferent: boolean
     fields: Partial<{
       holderName: string
       cardNumber: string
@@ -87,17 +88,11 @@ export type CustomOrganization = GetOrganizationQuery['getOrganizationByIdStoref
   users: GetOrganizationQuery['getUsers']
 }
 
-export type PaymentAddresType = {
-  addressId: { value: Address['addressId']; disabled?: boolean }
-  city: { value: Address['city']; disabled?: boolean }
-  complement: { value: Address['complement']; disabled?: boolean }
-  country: { value: Address['country']; disabled?: boolean }
-  neighborhood: { value: Address['neighborhood']; disabled?: boolean }
-  number: { value: Address['number']; disabled?: boolean }
-  postalCode: { value: Address['postalCode']; disabled?: boolean }
-  geoCoordinates: { value: Address['geoCoordinates']; disabled?: boolean }
-  receiverName: { value: Address['receiverName']; disabled?: boolean }
-  reference: { value: Address['reference']; disabled?: boolean }
-  state: { value: Address['state']; disabled?: boolean }
-  street: { value: Address['street']; disabled?: boolean }
+type PaymentField<T> = {
+  value: T
+  disabled?: boolean
+}
+
+export type PaymentAddressType = {
+  [K in keyof Address]: PaymentField<Address[K]>
 }