+++ react/utils/index.ts
@@ -1,6 +1,6 @@
-import { Item, Maybe, PaymentData } from 'vtex.checkout-graphql'
+import { Address, Item, PaymentData } from 'vtex.checkout-graphql'
 
-import { PaymentAddresType } from '../typings'
+import { PaymentAddressType } from '../typings'
 
 export * from './messages'
 
@@ -32,51 +32,18 @@ export function getFirstInstallmentByPaymentSystem(
   return installmentOption?.installments[0]
 }
 
-type AddressData = {
-  [key: string]: {
-    value:
-      | string
-      | number[]
-      | number
-      | null
-      | undefined
-      | Array<Maybe<number>>
-      | boolean
-    disabled?: boolean
-  }
-}
-
-type ExtractedValues = {
-  [key: string]:
-    | string
-    | number[]
-    | number
-    | null
-    | undefined
-    | Array<Maybe<number>>
-    | boolean
-}
-
-export function extractAddressValues(data: AddressData): ExtractedValues {
-  const extractedValues: ExtractedValues = {}
-
-  for (const key in data) {
-    if (data[key].value !== undefined && data[key].value !== null) {
-      extractedValues[key] = data[key].value
-    }
-  }
-
-  return extractedValues
-}
-
 export function toggleAddress(
-  data: PaymentAddresType,
+  data: PaymentAddressType,
   enabled: boolean
-): PaymentAddresType {
+): PaymentAddressType {
   const newAddresss = { ...data }
 
   for (const key in newAddresss) {
-    newAddresss[key as keyof PaymentAddresType].disabled = !enabled
+    const field = newAddresss[key as keyof PaymentAddressType]
+
+    if (field) {
+      field.disabled = !enabled
+    }
   }
 
   return newAddresss
@@ -100,3 +67,14 @@ export const getEmptyAddress = (country: string) => {
     addressQuery: null,
   }
 }
+
+export const buildBillingAddress = (
+  newAddress?: Address | null,
+  shippingAddress?: Address | null
+) => ({
+  ...newAddress,
+  addressId: newAddress?.addressId ?? '0',
+  addressQuery: null,
+  addressType: 'commercial',
+  receiverName: newAddress?.receiverName ?? shippingAddress?.receiverName ?? '',
+})