+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Using invoiceData of orderForm.
+
 ## [0.0.9] - 2024-11-19
 
 ### Added